+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # NEXT RELEASE
+
+### v0.70.3
 * Add `isLocal` getters to `Account` and `Toot` objects, improve `extractDomain()` helper method
 * Add `isLocalUrl()` method to MastoApi
-* Don't rebuild Account objects unnecessarily
-* Uniquify Account objects by url, not by ID (there can be duplicate IDs depending on server)
+* Don't rebuild `Account` objects unnecessarily
+* Uniquify `Account` objects by `url`, not by `id` (IDs are only unique across a single server)
 
 ### v0.70.2
 * Export `optionalSuffix()`